UD-12  [WIP] [feat]: created loading spinner and made remake on the dashboard page
