+++ src/app/admin/list/government/page.tsx
@@ -1,4 +1,5 @@
 "use client";
+import Governments from "@/components/lists/Governments";
 import {
 	AlertDialog,
 	AlertDialogAction,
@@ -42,53 +43,10 @@ import { toast } from "sonner";
 const pageListGovernment = () => {
 	const router = useRouter();
 	const [isOpen, setIsOpen] = useState(false);
-	const [isAlert, setIsAlert] = useState(false);
-	const [id, setId] = useState("");
-
-	const { data: government, refetch } = useQuery({
-		queryKey: ["get-government"],
-		queryFn: () => getGovernmentForm(),
-		// enabled: false,
-	});
-
-	const { mutateAsync: governmentDelete } = useMutation({
-		mutationKey: ["delete-government", id],
-		mutationFn: () => deleteGovernment(id),
-	});
-
 	const {
 		actions: { logout },
 		state: { user },
 	} = AuthStore();
-
-	const handleDelete = async () => {
-		const inviteForm = async () => {
-			try {
-				await governmentDelete();
-			} catch (error) {}
-		};
-
-		toast.promise(inviteForm, {
-			loading: "Carregando...",
-			duration: 4000,
-
-			success: () => {
-				router.back();
-				return "Sistema de Governo Removido";
-			},
-
-			error: "Erro ao remover o sistema de governo",
-
-			style: {
-				boxShadow: "1px 2px 20px 6px #555",
-			},
-		});
-	};
-
-	const handleClick = (id: string) => {
-		setId(id);
-		setIsAlert(true);
-	};
 
 	return (
 		<>
@@ -167,72 +125,7 @@ const pageListGovernment = () => {
 							</div>
 						</div>
 					</div>
-					{government?.length ? (
-						<>
-							{government?.map((item) => (
-								<div key={item.id} className="py-4">
-									<div className="grid grid-cols-party px-20 2xl:px-32 h-[75px] 2xl:h-[80px] items-center 2xl:text-xl mplus">
-										<div className="grid grid-cols-nameparty items-center">
-											<div className="flex justify-center" />
-
-											<div className="flex px-6">
-												<span className="truncate 2xl:font-medium text-lg">
-													{item.name}
-												</span>
-											</div>
-										</div>
-										<div className="flex justify-center">
-											<span className="pr-8 2xl:pr-2 2xl:font-medium text-lg">
-												{item.cod}
-											</span>
-										</div>
-										<div className="2xl:px-14">
-											<span className="truncate 2xl:font-medium text-lg">
-												{item.description}
-											</span>
-										</div>
-										<div className="flex justify-end">
-											<div className="flex justify-end items-center">
-												<DropdownMenu>
-													<DropdownMenuTrigger asChild>
-														<Button variant="ghost">
-															<EllipsisVertical className="h-[25px] w-[25px]" />
-														</Button>
-													</DropdownMenuTrigger>
-													<DropdownMenuContent className="w-20">
-														<DropdownMenuGroup>
-															<DropdownMenuItem
-																onClick={() =>
-																	router.push(
-																		`/admin/edit/${item.id}/government`,
-																	)
-																}
-															>
-																Editar
-															</DropdownMenuItem>
-															<DropdownMenuItem
-																className="text-red-500 focus:text-red-400"
-																onClick={() => handleClick(item.id)}
-															>
-																Remover
-															</DropdownMenuItem>
-														</DropdownMenuGroup>
-													</DropdownMenuContent>
-												</DropdownMenu>
-											</div>
-										</div>
-									</div>
-								</div>
-							))}
-						</>
-					) : (
-						<div className="w-full py-32 flex justify-center">
-							<p className="text-2xl">
-								Infelizmente não foi encontrado nenhum resultado para a sua
-								busca!
-							</p>
-						</div>
-					)}
+					<Governments />
 				</div>
 			</main>
 			<Sheet open={isOpen} onOpenChange={setIsOpen}>
@@ -262,30 +155,6 @@ const pageListGovernment = () => {
 					</div>
 				</SheetContent>
 			</Sheet>
-			<AlertDialog open={isAlert} onOpenChange={setIsAlert}>
-				<AlertDialogContent className="bg-white">
-					<AlertDialogHeader>
-						<AlertDialogTitle>
-							Você realmente tem certeza disso?
-						</AlertDialogTitle>
-						<AlertDialogDescription>
-							Você está prestes a remover um sistema de governo. Deseja
-							realmente continuar?
-						</AlertDialogDescription>
-					</AlertDialogHeader>
-					<AlertDialogFooter>
-						<AlertDialogCancel>Cancelar</AlertDialogCancel>
-						<AlertDialogAction
-							onClick={() => {
-								setIsAlert(false);
-								handleDelete();
-							}}
-						>
-							Continuar
-						</AlertDialogAction>
-					</AlertDialogFooter>
-				</AlertDialogContent>
-			</AlertDialog>
 		</>
 	);
 };
