+++ src/app/admin/list/voter/page.tsx
@@ -1,23 +1,8 @@
 "use client";
-import {
-	AlertDialog,
-	AlertDialogAction,
-	AlertDialogCancel,
-	AlertDialogContent,
-	AlertDialogDescription,
-	AlertDialogFooter,
-	AlertDialogHeader,
-	AlertDialogTitle,
-} from "@/components/ui/alert-dialog";
+import Voters from "@/components/lists/Voters";
+import Return from "@/components/main/Return";
 /* eslint-disable react-hooks/rules-of-hooks */
 import { Button } from "@/components/ui/button";
-import {
-	DropdownMenu,
-	DropdownMenuContent,
-	DropdownMenuGroup,
-	DropdownMenuItem,
-	DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
 import { Input } from "@/components/ui/input";
 import {
 	Sheet,
@@ -26,70 +11,21 @@ import {
 	SheetTitle,
 } from "@/components/ui/sheet";
 import filter from "@/img/filter.svg";
-import iconBack from "@/img/icon-back.svg";
+
 import logoIf from "@/img/logo-if.svg";
-import { deleteVoter } from "@/requests/voter/delete";
-import { getVoters } from "@/requests/voter/findAll";
 import { AuthStore } from "@/store/auth";
-import { useMutation, useQuery } from "@tanstack/react-query";
-import { CirclePlus, EllipsisVertical, UserRound } from "lucide-react";
+import { CirclePlus, UserRound } from "lucide-react";
 import Image from "next/image";
 import Link from "next/link";
-import { useRouter } from "next/navigation";
-import router from "next/router";
 import { useState } from "react";
-import { toast } from "sonner";
 
 const pageListVoter = () => {
-	const router = useRouter();
 	const [isOpen, setIsOpen] = useState(false);
-	const [isAlert, setIsAlert] = useState(false);
-	const [id, setId] = useState("");
 
 	const {
 		actions: { logout },
 		state: { user },
 	} = AuthStore();
-
-	const { data: voters } = useQuery({
-		queryKey: ["get voter"],
-		queryFn: getVoters,
-	});
-
-	const { mutateAsync: voterDelete } = useMutation({
-		mutationKey: ["delete-voter", id],
-		mutationFn: () => deleteVoter(id),
-	});
-
-	const handleClick = (id: string) => {
-		setId(id);
-		setIsAlert(true);
-	};
-
-	const handleDelete = async () => {
-		const inviteForm = async () => {
-			try {
-				await voterDelete();
-			} catch (error) {}
-		};
-
-		toast.promise(inviteForm, {
-			loading: "Carregando...",
-			duration: 4000,
-
-			success: () => {
-				router.back();
-				return "Partido Removido";
-			},
-
-			error: "Erro ao remover o partido",
-
-			style: {
-				boxShadow: "1px 2px 20px 6px #555",
-			},
-		});
-	};
-
 	return (
 		<>
 			<main className="grid grid-cols-10 mx-auto min-h-screen">
@@ -117,19 +53,7 @@ const pageListVoter = () => {
 					</div>
 					{/* NAVBAR */}
 					<div className="grid grid-cols-4 2xl:px-6">
-						<div className="flex items-center px-5">
-							<Button
-								className="hover:bg-transparent"
-								variant="ghost"
-								onClick={() => router.back()}
-							>
-								<Image
-									className="h-12 2xl:h-14 2xl:w-14 w-12"
-									src={iconBack}
-									alt="Ícone voltar"
-								/>
-							</Button>
-						</div>
+						<Return />
 						<div className="col-span-3 flex items-center px-4 2xl:px-14">
 							<Button className="hover:bg-transparent" variant="ghost">
 								<Image
@@ -170,69 +94,8 @@ const pageListVoter = () => {
 							</div>
 						</div>
 					</div>
-					{voters?.length ? (
-						<>
-							{voters?.map((item) => (
-								<div key={item.id} className="py-4">
-									<div className="mplus 2xl:space-x-2 2xl:text-lg 2xl:font-medium grid grid-cols-party px-20 2xl:px-32 h-[75px] 2xl:h-[80px] items-center">
-										<div className="grid grid-cols-nameparty items-center">
-											<div className="flex justify-center" />
-
-											<div className="px-6 2xl:px-7">
-												<span className="truncate text-[#121212]">
-													{item.name}
-												</span>
-											</div>
-										</div>
-										<div className="px-11 2xl:px-14">
-											<span className="">{item.enrollment}</span>
-										</div>
-										<div className="px-16 2xl:px-[4.55rem]">
-											<span className="truncate">{item.class}</span>
-										</div>
-										<div className="grid grid-cols-2">
-											<div className="px-5 2xl:px-9">
-												<span className="truncate">{item.email}</span>
-											</div>
-											<div className="flex justify-end items-center">
-												<DropdownMenu>
-													<DropdownMenuTrigger asChild>
-														<Button variant="ghost">
-															<EllipsisVertical className="h-[25px] w-[25px]" />
-														</Button>
-													</DropdownMenuTrigger>
-													<DropdownMenuContent className="w-20">
-														<DropdownMenuGroup>
-															<DropdownMenuItem
-																onClick={() =>
-																	router.push(`/admin/edit/${item.id}/voter`)
-																}
-															>
-																Editar
-															</DropdownMenuItem>
-															<DropdownMenuItem
-																className="text-red-500 focus:text-red-400"
-																onClick={() => handleClick(item.id)}
-															>
-																Remover
-															</DropdownMenuItem>
-														</DropdownMenuGroup>
-													</DropdownMenuContent>
-												</DropdownMenu>
-											</div>
-										</div>
-									</div>
-								</div>
-							))}
-						</>
-					) : (
-						<div className="w-full py-32 flex justify-center">
-							<p className="text-2xl">
-								Infelizmente não foi encontrado nenhum resultado para a sua
-								busca!
-							</p>
-						</div>
-					)}
+					{/* Eleitores */}
+					<Voters />
 				</div>
 			</main>
 			<Sheet open={isOpen} onOpenChange={setIsOpen}>
@@ -262,30 +125,6 @@ const pageListVoter = () => {
 					</div>
 				</SheetContent>
 			</Sheet>
-			<AlertDialog open={isAlert} onOpenChange={setIsAlert}>
-				<AlertDialogContent className="bg-white">
-					<AlertDialogHeader>
-						<AlertDialogTitle>
-							Você realmente tem certeza disso?
-						</AlertDialogTitle>
-						<AlertDialogDescription>
-							Você está prestes a remover um Eleitor. Deseja realmente
-							continuar?
-						</AlertDialogDescription>
-					</AlertDialogHeader>
-					<AlertDialogFooter>
-						<AlertDialogCancel>Cancelar</AlertDialogCancel>
-						<AlertDialogAction
-							onClick={() => {
-								setIsAlert(false);
-								handleDelete();
-							}}
-						>
-							Continuar
-						</AlertDialogAction>
-					</AlertDialogFooter>
-				</AlertDialogContent>
-			</AlertDialog>
 		</>
 	);
 };
