+++ src/requests/government/delete.ts
@@ -1,5 +1,7 @@
 import { api } from "../api";
 
 export async function deleteGovernment(id: string) {
-	await api.delete(`/government/form/${id}`);
+	const response = await api.delete(`/government/form/${id}`);
+
+	return { response };
 }
