+++ src/requests/politicalPart/delete.ts
@@ -1,5 +1,7 @@
 import { api } from "../api";
 
 export async function deletePoliticalParty(idPoliticalParty: string) {
-	await api.delete(`/political/${idPoliticalParty}`);
+	const response = await api.delete(`/political/${idPoliticalParty}`);
+
+	return { response };
 }
