+++ src/requests/voter/delete.ts
@@ -1,5 +1,6 @@
 import { api } from "../api";
 
 export async function deleteVoter(idVoter: string) {
-	await api.delete(`/voter/${idVoter}`);
+	const response = await api.delete(`/voter/${idVoter}`);
+	return { response };
 }
